+++ packages/banana-react/src/index.ts
@@ -5,8 +5,10 @@ import { Countdown } from './countdown';
 import { Divider } from './divider';
 import { Dropdown } from './dropdown';
 import { Input } from './input';
+import { Marquee } from './marquee';
 import { Menu } from './menu';
 import { MenuItem } from './menu-item';
+import { Message } from './message';
 import { Overlay } from './overlay';
 import { Popup } from './popup';
 import { Progress } from './progress';
@@ -22,8 +24,10 @@ export {
   Divider,
   Dropdown,
   Input,
+  Marquee,
   Menu,
   MenuItem,
+  Message,
   Overlay,
   Popup,
   Progress,